01_etl/postgres_to_es/requirements.txt
pydantic==1.9.0
django-environ==0.8.1
psycopg2-binary==2.9.3
redis==4.2.2
elasticsearch==7.17.3